packfile - imcc
git-svn-id: https://svn.parrot.org/parrot/trunk@2970 d31e2699-5ff4-0310-a27c-f18f2fbe73fe

+++ languages/imcc/imc.h
@@ -78,7 +78,7 @@ EXTERN SymReg** reglist;
 
 EXTERN char * sourcefile;	/* current file */
 EXTERN char * function;	/* current function */
-EXTERN long        line;	/* and line */
+EXTERN int        line;	/* and line */
 EXTERN char optimizer_opt[20];
 EXTERN int dont_optimize;
